Update start.bat
Lots of changes. Chances are, they won't all be listed here due to the amount of code that was rewritten and changed over time - I'll have to get better about documenting this. Biggest change is that .vbs scripts are now entirely temporary. The code is echoed and forwarded to a new temporary file which is renamed to TMPxxx.vbs and run, before being deleted. This cleans up the main folder significantly.